[Feature] Player Abstraction & Character Classes
Created rogue, mage and berserker classes; made player an abstract class and made each class scale their damage uniquely.
